+++ app/Http/Controllers/ImportController.php
@@ -5,14 +5,14 @@ declare(strict_types=1);
 namespace Kami\Cocktail\Http\Controllers;
 
 use Throwable;
-use Symfony\Component\Yaml\Yaml;
 use OpenApi\Attributes as OAT;
-use Kami\Cocktail\OpenAPI as BAO;
 use Illuminate\Http\JsonResponse;
+use Kami\Cocktail\OpenAPI as BAO;
 use Kami\Cocktail\Models\Cocktail;
 use Kami\Cocktail\Scraper\Manager;
-use Kami\Cocktail\External\Import\FromArray;
+use Kami\Cocktail\External\Import\FromSchemaDraft2;
 use Kami\Cocktail\Http\Requests\ImportRequest;
+use Kami\Cocktail\Http\Requests\ScrapeRequest;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Kami\Cocktail\Http\Resources\CocktailResource;
 use Kami\Cocktail\External\Import\DuplicateActionsEnum;
@@ -21,14 +21,12 @@ class ImportController extends Controller
 {
     #[OAT\Post(path: '/import/cocktail', tags: ['Import'], summary: 'Import a cocktail', parameters: [
         new BAO\Parameters\BarIdParameter(),
-        new OAT\Parameter(name: 'type', in: 'query', description: 'Type of import', required: true, schema: new OAT\Schema(type: 'string', enum: ['url', 'json', 'yaml', 'yml', 'collection'])),
-        new OAT\Parameter(name: 'save', in: 'query', description: 'Save imported cocktails to the database', schema: new OAT\Schema(type: 'boolean')),
     ], requestBody: new OAT\RequestBody(
         required: true,
         content: [
             new OAT\JsonContent(type: 'object', properties: [
-                new OAT\Property(property: 'source', type: 'string', example: 'https://www.example.com/recipe-url'),
-                new OAT\Property(property: 'duplicate_actions', ref: DuplicateActionsEnum::class, example: '0'),
+                new OAT\Property(property: 'schema', ref: 'https://raw.githubusercontent.com/karlomikus/bar-assistant/v4dev/docs/cocktail-02.schema.json'),
+                new OAT\Property(property: 'duplicate_actions', ref: DuplicateActionsEnum::class, example: 'none', description: 'How to handle duplicates. Cocktails are matched by lowercase name.'),
             ]),
         ]
     ))]
@@ -36,57 +34,54 @@ class ImportController extends Controller
         new BAO\WrapObjectWithData(BAO\Schemas\Cocktail::class),
     ])]
     #[BAO\NotAuthorizedResponse]
-    public function cocktail(ImportRequest $request, FromArray $arrayImporter): JsonResponse|JsonResource
+    public function cocktail(ImportRequest $request, FromSchemaDraft2 $importer): JsonResource
     {
         if ($request->user()->cannot('create', Cocktail::class)) {
             abort(403);
         }
 
-        $dataToImport = [];
-        $type = $request->get('type', 'url');
-        $save = $request->get('save', false);
-        $source = $request->post('source');
-        $duplicateAction = DuplicateActionsEnum::from((int) $request->post('duplicate_actions', '0'));
+        $source = $request->post('schema');
+        $duplicateAction = DuplicateActionsEnum::from($request->post('duplicate_actions', 'none'));
 
-        if ($type === 'url') {
-            $request->validate(['source' => 'url']);
-            try {
-                $scraper = Manager::scrape($source);
-            } catch (Throwable $e) {
-                abort(400, $e->getMessage());
-            }
+        $cocktail = $importer->process($source, $request->user()->id, bar()->id, $duplicateAction);
 
-            $dataToImport = $scraper->toArray();
+        return new CocktailResource($cocktail);
+    }
+
+    #[OAT\Post(path: '/import/scrape', tags: ['Import'], summary: 'Scrape a recipe', description: 'Try to scrape a recipe from a website. Most of the well known recipe websites should work. Data returned is a valid JSON schema that you can import using import cocktail endpoint.', parameters: [
+        new BAO\Parameters\BarIdParameter(),
+    ], requestBody: new OAT\RequestBody(
+        required: true,
+        content: [
+            new OAT\JsonContent(type: 'object', properties: [
+                new OAT\Property(property: 'source', type: 'string', example: 'https://www.example.com/recipe-url'),
+            ]),
+        ]
+    ))]
+    #[OAT\Response(response: 200, description: 'Successful response', content: [
+        new OAT\JsonContent(type: 'object', properties: [
+            new OAT\Property(property: 'schema_version', type: 'string', example: 'draft2'),
+            new OAT\Property(property: 'schema', ref: 'https://raw.githubusercontent.com/karlomikus/bar-assistant/v4dev/docs/cocktail-02.schema.json'),
+            new OAT\Property(property: 'scraper_meta', type: 'object', additionalProperties: true),
+        ]),
+    ])]
+    #[BAO\NotAuthorizedResponse]
+    public function scrape(ScrapeRequest $request): JsonResponse|JsonResource
+    {
+        if ($request->user()->cannot('create', Cocktail::class)) {
+            abort(403);
         }
 
-        if ($type === 'json') {
-            if (!is_array($source)) {
-                if (!$source = json_decode($source, true)) {
-                    abort(400, 'Unable to parse the JSON string');
-                }
-            }
-
-            $dataToImport = $source;
+        try {
+            $scraper = Manager::scrape($request->post('source'));
+        } catch (Throwable $e) {
+            abort(400, $e->getMessage());
         }
 
-        if ($type === 'yaml' || $type === 'yml') {
-            try {
-                $dataToImport = Yaml::parse($source);
-            } catch (Throwable) {
-                abort(400, sprintf('Unable to parse the YAML string'));
-            }
-        }
-
-        if ($save) {
-            $cocktail = $arrayImporter->process($dataToImport, $request->user()->id, bar()->id);
-            $cocktail->load(['ingredients.ingredient', 'images' => function ($query) {
-                $query->orderBy('sort');
-            }, 'tags', 'glass', 'ingredients.substitutes', 'method', 'createdUser', 'updatedUser', 'collections', 'utensils']);
-            $dataToImport = new CocktailResource($cocktail);
-        }
+        $dataToImport = $scraper->toArray();
 
         return response()->json([
-            'data' => $dataToImport
+            'data' => $dataToImport,
         ]);
     }
 }